Add TaskButtonFactory.kt
