+++ ground/src/main/java/com/google/android/ground/ui/datacollection/components/TaskButton.kt
@@ -15,17 +15,8 @@
  */
 package com.google.android.ground.ui.datacollection.components
 
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
-import android.widget.Button
-import android.widget.ImageButton
-import androidx.core.content.res.ResourcesCompat
-import com.google.android.ground.databinding.*
 import com.google.android.ground.model.submission.TaskData
-import com.google.android.ground.ui.datacollection.components.ButtonAction.Theme
-import com.google.android.ground.ui.datacollection.components.ButtonAction.Type
-import com.google.android.material.button.MaterialButton
 
 /** Wrapper class for holding a button. */
 data class TaskButton(private val view: View) {
@@ -78,64 +69,4 @@ data class TaskButton(private val view: View) {
   fun onTaskDataUpdated(taskData: TaskData?) {
     taskUpdatedCallback?.let { it(this, taskData) }
   }
-
-  companion object {
-    /** Inflates the button layout and attaches to the given container view. */
-    fun createAndAttachButton(
-      action: ButtonAction,
-      container: ViewGroup,
-      layoutInflater: LayoutInflater
-    ) =
-      TaskButton(
-        when (action.type) {
-          Type.TEXT -> createTextTypeButtons(layoutInflater, container, action)
-          Type.ICON -> createIconTypeButtons(layoutInflater, container, action)
-          Type.TEXT_ICON -> createTextIconTypeButtons(layoutInflater, container, action)
-        }
-      )
-  }
 }
-
-// TODO(Shobhit): Figure out a way to create styled buttons without using XML.
-private fun createTextTypeButtons(
-  layoutInflater: LayoutInflater,
-  container: ViewGroup,
-  action: ButtonAction
-): Button =
-  when (action.theme) {
-    Theme.DARK_GREEN -> TaskChipButtonDarkGreenBinding.inflate(layoutInflater, container).button
-    Theme.LIGHT_GREEN -> TaskChipButtonLightGreenBinding.inflate(layoutInflater, container).button
-    Theme.OUTLINED -> TaskChipButtonOutlineBinding.inflate(layoutInflater, container).button
-    Theme.TRANSPARENT -> TaskChipButtonTransparentBinding.inflate(layoutInflater, container).button
-  }.apply { action.textId?.let { setText(it) } }
-
-// TODO(Shobhit): Figure out a way to create styled buttons without using XML.
-private fun createIconTypeButtons(
-  layoutInflater: LayoutInflater,
-  container: ViewGroup,
-  action: ButtonAction
-): ImageButton =
-  if (action.theme == Theme.LIGHT_GREEN) {
-      TaskChipButtonWithIconLightGreenBinding.inflate(layoutInflater, container).button
-    } else {
-      error("Unsupported icon type button for theme: ${action.theme}")
-    }
-    .apply {
-      action.drawableId?.let { setImageDrawable(ResourcesCompat.getDrawable(resources, it, null)) }
-    }
-
-// TODO(Shobhit): Figure out a way to create styled buttons without using XML.
-private fun createTextIconTypeButtons(
-  layoutInflater: LayoutInflater,
-  container: ViewGroup,
-  action: ButtonAction
-): MaterialButton =
-  if (action.theme == Theme.OUTLINED) {
-      TaskChipButtonWithTextAndIconTransparentBinding.inflate(layoutInflater, container).button
-    } else {
-      error("Unsupported icon type button for theme: $action.theme")
-    }
-    .apply {
-      action.textId?.let { setText(it) }
-      action.drawableId?.let { icon = ResourcesCompat.getDrawable(resources, it, null) }
-    }
